rename constants file
